uni_crawler_backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
uvloop==0.19.0
celery[redis]==5.3.6
redis==5.0.3
playwright==1.42.0
httpx==0.27.0
fake-useragent==1.5.1
python-dotenv==1.0.1
websockets==12.0
beautifulsoup4==4.12.3  # Fallback parsing
nltk==3.8.1             # For simple tokenization/NLP
asgiref==3.7.2
selectolax==0.3.21
pypdf==4.0.1